ARM: mvebu: rename mvebu_defconfig to mvebu_v7_defconfig
With the merge of the Kirkwood support into arch/arm/mach-mvebu/, the
mvebu platform will no longer only contain ARMv7 compatible processors
(Armada 370, Armada XP, and soon Armada 375, Armada 38x and Dove), but
also ARMv5 compatible processors (Kirkwood, and hopefully Orion5x in
the future).

However, a single mvebu_defconfig cannot work, since it is not
possible to build a kernel that supports both ARMv5 and ARMv7
platforms in the same binary. As a consequence, this commit renames
mvebu_defconfig to mvebu_v7_defconfig, which is the configuration that
will build a kernel that supports all ARMv7 mvebu platforms. As
Kirkwood support gets merged into mach-mvebu, an additional
mvebu_v5_defconfig will be added.

Even though we already have a multi_v7_defconfig, the mvebu developers
found it more convenient for development to have a defconfig that
builds only the mvebu platforms.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Acked-by: Andrew Lunn <andrew@lunn.ch>
Signed-off-by: Jason Cooper <jason@lakedaemon.net>