X.509: Remove obj path in certificate file for MOD_SIG
Remove $(objpath) path reference for signing_key.x509
file such that it remains the same whether the file
exists or not.

CONFIG_MOD_SIG includes the signing_key.x509 file
that gets embedded into the image, with .x509.list
tracking the list of certificates.  If signing_key.x509
does not exist, .x509.list will contain a path
reference ("./signing_key.x509") but on subsequent
compiles the path reference is stripped.  This leads to
relinking of vmlinux.

Change-Id: I3e3fefe29afbb3da7542fc16ce662177e41c3562
Signed-off-by: David Ng <dave@codeaurora.org>