input: misc: fpc1145: Unconditionally enable irq wake on suspend.
The current polling state has a tendency to run into race conditions
with device suspend, where the irq_fired flag may have not been cleared
before a suspend is requested.
This is particularly evident when a finger is still on the sensor; the
HAL waits for it to disappear while the wakelock times out and fires off
the suspend. Which in itself is another issue that is resolved in future
commits.
Furthermore, irq_wake is disabled unconditionally, causing "unbalanced
disable" warnings every so often.

Change-Id: Icaf2b5f77ead7def06f60b6ba3d852a58d335dec
Signed-off-by: MarijnS95 <marijns95@gmail.com>